requirements.txt
aiosqlite==0.10.0
asyncpg==0.18.3
ciso8601==2.1.1
Click==7.0
fastapi==0.38.1
gunicorn==19.9.0
h11==0.8.1
httptools==0.0.13
pydantic==0.32.2
PyPika==0.35.8
starlette==0.12.8
tortoise-orm==0.13.9
uvicorn==0.9.0
uvloop==0.13.0
websockets==8.0.2